+++ tests/test_views.py
@@ -34,31 +34,10 @@ def test_pagination_view() -> None:
     class PaginationView(web.View, PaginationMixin):
         page_adapter = int
 
-    pk = '1'
-    req = make_mocked_request(METH_GET, f'/?page=2')
+    req = make_mocked_request(METH_GET, '/?page=2')
     view = PaginationView(req)
     assert isinstance(view.page, int)
     assert view.page == 2
-
-
-def test_integer_pk_view_deprecated() -> None:
-    class IntegerPrimaryKeyView(web.View, PrimaryKeyMixin):
-        pk_factory = int
-
-    pk = '1'
-    req = make_mocked_request(METH_GET, f'/{pk}', match_info={'pk': pk})
-    view = IntegerPrimaryKeyView(req)
-    assert isinstance(view.pk, int)
-
-
-def test_uuid_pk_view_deprecated() -> None:
-    class UUIDPrimaryKeyView(web.View, PrimaryKeyMixin):
-        pk_factory = uuid.UUID
-
-    pk = str(uuid.uuid4())
-    req = make_mocked_request(METH_GET, f'/{pk}', match_info={'pk': pk})
-    view = UUIDPrimaryKeyView(req)
-    assert isinstance(view.pk, uuid.UUID)
 
 
 def test_integer_pk_view() -> None: